Tutorial progress indicator refactor and add tests.
This refactors the code to use localStorage for caching, adds a new custom hook useTutorialList, which and makes useProgressIndicator a wrapper around that. The idea being we might want to use a tutorial list + caching to store additional types of personalized meta-data in the future.

The withDrupalOauthProvider HOC was updated to include code which can take an OAuth token and retrieve the Drupal UUID of a logged in user. Then store that in local state. This allows us to know not only that a user is logged in, but specifically which user. Which is necessary to ensure we purge the cache and not leak cached data between users.

Code was added to gatsby-browser.js to handle initialization of the DrupalOauth client and DrupalOauthContext.Provider so that it's done automatically and no longer needs to be done in the consuming applications Layout.jsx or similar. This also means consuming apps will need to be updated to remove that now uncessary code.

Adds tests for the new useProgressIndicator, and useTutorialList hooks, and updates other tests as needed.