Allow use of cached response when retrieving list of tutorials.
When the user is anon do not include the tutorial_read_state field in the query for a list of tutorials. This allows the response to be cached. On the Drupal side the tutorial_read_state field is non-cachable, so including it in the request means the response can never be cached.