Allow use of cached response when retrieving a list of tutorials from Drupal.
On the Drupal side the tutorial_read_state field is not cachable. So when it's included in the API request the returned response is not cacheable. However, anon users don't need the tutorial_read_state data because we don't track read state for anon users. So, by changing the API request for anon users to NOT include the field the resulting response can be cached by Drupal, CDNs and the user's browser. Yay!