Do not ping API for personilization data for anon users.
Right now we ping Drupal for a list of tutorial read state values even if the user is not signed in. But we don't need the data, and it's causing at least two requestions to Pantheon for every anon. page load. An OPTIONS request and a GET request. But the data it retrieves is never used for anon users. So we don't need to get it.